Modified DMTF threshold logs to include normal range
In the DMTF Registry, there's a message that indicates the sensor is out
of the warning and critical ranges. It says the sensor is back in the
normal range. This change adds logic to check if a sensor is in the
normal range so we don't say it's still outside the normal range
whenever a sensor leaves the fatal/critical/warning range.

Signed-off-by: Brooks Keough <bkeough@nvidia.com>

Fixes NVBug https://nvbugspro.nvidia.com/bug/5121116