Restore boost headers to fix build
Due to build option defaults mismatching between the meson.options file
and the Yocto recipe, the CI is building a different set of default code
than what is built by Yocto.

Because of this, removing the boost headers worked fine in CI but is
failing in Yocto, so restore the boost headers to fix the default Yocto
build until we can remove the boost dependency there as well.

Change-Id: I85f7d44d197dd67cdafa7c10736179adcd0efa33
Signed-off-by: Jason M. Bills <jason.m.bills@linux.intel.com>